lstm bptt

+++ src/main/java/ohs/ml/eval/Metrics.java
@@ -10,11 +10,11 @@ public class Metrics {
 		return ret;
 	}
 
-	public static double f1(int correct_cnt, int answer_cnt, int predict_cnt) {
+	public static double f1(double correct_cnt, double answer_cnt, double predict_cnt) {
 		return f1(precision(correct_cnt, predict_cnt), recall(correct_cnt, answer_cnt));
 	}
 
-	public static double precision(int correct_cnt, int predict_cnt) {
+	public static double precision(double correct_cnt, double predict_cnt) {
 		double ret = 0;
 		if (predict_cnt > 0) {
 			ret = 1f * correct_cnt / predict_cnt;
@@ -22,7 +22,7 @@ public class Metrics {
 		return ret;
 	}
 
-	public static double recall(int correct_cnt, int answer_cnt) {
+	public static double recall(double correct_cnt, double answer_cnt) {
 		double ret = 0;
 		if (answer_cnt > 0) {
 			ret = 1f * correct_cnt / answer_cnt;
